CE-355 OSPF Unnumbered 2s + 2s2l
Merge the 2s & 2s2l "topology" tests together (everything is selected out
from JSON anyway)
Move the interfaces.rb tests to the OSPF tests as they're really specific
to the OSPF workbench configuration.
Move the resolver.rb test to "core" as it's nothing to do with OSPF and is a
good core test; it only relies on knowing the names of the hosts for a given
topology, and those don't change.
Remove the 2s & 2s2l test directories.