Remove the explicit topology from the rake target
Make the Rake targets nuetral E.g. 'core' rather than 'core2s', 'core2s2l'
etc. and allow the user to pass the topology as an argument to Rake.
Note the additional Rake hackery; Rake supports passing arguments to a task
but the syntax for it is foul, and this is slightly more user friendly even if
it is a giant hack.