CE-541 Tweak ruby-lint testing
I'm not convinced that invoking ruby-lint was doing what we wanted, so I've
tweaked how it's invoked (once per. file in a loop).
Moved the settings out to ruby-lint.yml and disabled checks for
undefined_methods as it was triggering on all of the Serverspec DSL; there's
no definition for Serverspec in ruby-lint so that's the best we can do.